Remove long double from DataIterator

+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/datatables/GetArrayByIndex.java
@@ -61,8 +61,7 @@ public class GetArrayByIndex extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         BDataTable dataTable = (BDataTable) getArgument(ctx, 0);
-        //Here we cast index value to int as anyway java doesn't support long as array indexes
-        int index = (int) ((BInteger) getArgument(ctx, 1)).intValue();
+        long index = ((BInteger) getArgument(ctx, 1)).intValue();
         Map<String, Object> arrayMap = dataTable.getArray(index);
         BMap<BString, BValue> returnMap = new BMap<>();
         if (arrayMap != null && !arrayMap.isEmpty()) {
