+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/datatables/GetFloatByIndex.java
@@ -56,8 +56,7 @@ public class GetFloatByIndex extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         BDataTable dataTable = (BDataTable) getArgument(ctx, 0);
-        //Here we cast index value to int for consistency
-        int index = (int) ((BInteger) getArgument(ctx, 1)).intValue();
+        long index = ((BInteger) getArgument(ctx, 1)).intValue();
         return getBValues(new BFloat(dataTable.getFloat(index)));
     }
 }
